requirements.txt
django[argon2]
psycopg2
djangorestframework
djoser
pylibmc
gunicorn
django-debug-toolbar
django-filter